* commit '9480a384a6739bc2f4d2a047da77959180f9305c':
  PTSD-61 Fixed frameRate that was accidentally stored as integral number
  PTSD-61 Updated BatchData iterators
  PTSD-61 Added dependant defaults to automatically use gpu/host implementations based off sysconfig.analysisHardware
  PTSD-61 Fixed minor bug in PCIe reporting.

+++ src/basecaller/appModules/TraceFileDataSource.h
@@ -130,7 +130,7 @@ private:
     size_t currZmw_;
     size_t maxQueueSize_;
 
-    size_t frameRate_;
+    float frameRate_;
 
     std::string filename_;
